% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.matrix.mrf_penalty}
\alias{as.matrix.mrf_penalty}
\title{Convert a MRF penalty object to a matrix}
\usage{
\method{as.matrix}{mrf_penalty}(x, ...)
}
\arguments{
\item{x}{an object inheriting from class \code{"mrf_penalty"}}

\item{...}{arguments passed to other methods}
}
\description{
Convert a MRF penalty object to a matrix
}
\examples{
p <- mrf_penalty(1:10)
as.matrix(p)
}
