% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_config.R
\name{as_mrf_penalty}
\alias{as_mrf_penalty}
\title{Create a \code{"mrf_penalty"} object from a matrix and a config}
\usage{
as_mrf_penalty(penalty, config)
}
\arguments{
\item{penalty}{matrix}

\item{config}{list}
}
\description{
Create a \code{"mrf_penalty"} object from a matrix and a config
}
