% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{calc_subprec}
\alias{calc_subprec}
\title{Calculate precision matrix for a subset of nodes}
\usage{
calc_subprec(mat, indices)
}
\arguments{
\item{mat}{n x n positive semidefinite precision matrix, generated by the
Matrix package}

\item{indices}{numeric indices for the rows/columns of the submatrix to be
generated. All indices must be between 1 and n with no  repeated indices}
}
\description{
Internal function to find the precision matrix of a subset of
values (specified via indices) using block-inversion block-inversion of
sparse matrices to find the inverse of the submatrix for the selected
indices. Note that, even when given a sparse matrix, the precision matrix
for a subset of nodes will be in general dense.
}
\keyword{internal}
