% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_delta}
\alias{check_delta}
\title{Check delta parameter}
\usage{
check_delta(delta)
}
\description{
Internal function to check whether the delta parameter is valid
when added to the diagonal of a precision matrix
}
\keyword{internal}
