% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_config}
\alias{get_config}
\title{Extract configuration details of an MRF penalty}
\usage{
get_config(penalty)
}
\arguments{
\item{penalty}{an object of class \code{"mrf_penalty"}}
}
\value{
An object of class \code{"mrf_config"}, a list.
}
\description{
Extract configuration details of an MRF penalty
}
