% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_labels}
\alias{get_labels}
\title{Extract MRF node labels from an MRF penalty}
\usage{
get_labels(penalty)
}
\arguments{
\item{penalty}{an object of class \code{"mrf_penalty"}}
}
\description{
Extract MRF node labels from an MRF penalty
}
