% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model}
\alias{get_model}
\title{Extract the model type and parameters from an MRF penalty}
\usage{
get_model(penalty)
}
\arguments{
\item{penalty}{an object of class \code{"mrf_penalty"}}
}
\description{
Extract the model type and parameters from an MRF penalty
}
