% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mrf}
\alias{get_mrf}
\alias{get_mrf.bam}
\alias{get_mrf.gamm}
\alias{get_mrf.list}
\alias{get_mrf.gam}
\title{Extract a fitted MRF}
\usage{
get_mrf(object, ...)

\method{get_mrf}{bam}(object, ...)

\method{get_mrf}{gamm}(object, ...)

\method{get_mrf}{list}(object, ...)

\method{get_mrf}{gam}(object, term, ...)
}
\arguments{
\item{object}{An object from which to extract the fitted MRF. Currently only
for objects of classes \code{gam}, \code{bam}, and \code{gamm}, and GAMMs fitted by
\code{\link[gamm4:gamm4]{gamm4::gamm4()}}.}

\item{...}{Arguments passed to other methods.}

\item{term}{character; the MRF term to extract. Can be a partial match to a
term, which is matched against the smooth label.}
}
\value{
A object representing the fitted MRF
}
\description{
Extract a fitted MRF
}
