% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_obj}
\alias{get_obj}
\alias{get_obj.mrf_config}
\alias{get_obj.mrf_penalty}
\title{Extract the original object used to construct an MRF penalty}
\usage{
get_obj(object)

\method{get_obj}{mrf_config}(object)

\method{get_obj}{mrf_penalty}(object)
}
\arguments{
\item{object}{an object of class \code{"mrf_penalty"} or \code{"mrf_config"}.}
}
\value{
A length 1 character vector containing the type of MRF penalty.
}
\description{
Extract the original object used to construct an MRF penalty
}
