% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{get_treedist}
\alias{get_treedist}
\title{Path length distance between two nodes of a phylogenetic tree}
\usage{
get_treedist(tree, tip1, tip2)
}
\arguments{
\item{tree}{a phylo4 tree object}

\item{tip1}{The starting tip (either named or as an index)}

\item{tip2}{the ending tip for the path}
}
\description{
Internal function to get the shortest phylogenetic distance
between two nodes in a given tree
}
\keyword{internal}
