% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_type}
\alias{get_type}
\title{Extract the type of MRF from the penalty}
\usage{
get_type(penalty)
}
\arguments{
\item{penalty}{an object of class \code{"mrf_penalty"}}
}
\value{
A length 1 character vector containing the type of MRF penalty.
}
\description{
Extract the type of MRF from the penalty
}
