% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_config.R
\name{mrf_config}
\alias{mrf_config}
\title{MRF penalty configuration data}
\usage{
mrf_config(
  type = NULL,
  model = NULL,
  params = NULL,
  node_labels = NULL,
  delta = NULL,
  obj = NULL
)
}
\arguments{
\item{type}{character}

\item{model}{character}

\item{params}{list?}

\item{node_labels}{character}

\item{delta}{numeric}

\item{obj}{character}
}
\description{
MRF penalty configuration data
}
