% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{mrf_penalty.factor}
\alias{mrf_penalty.factor}
\title{Fully connected graph and random effect MRF penalties from a factor}
\usage{
\method{mrf_penalty}{factor}(
  object,
  type = c("full", "individual"),
  node_labels = NULL,
  add_delta = FALSE,
  ...
)
}
\arguments{
\item{object}{an R object to create the MRF penalty from.}

\item{type}{character; one of \code{"full"} or \code{"individual"} indicating if a
fully connected graph (\code{"full"}) or a random effect (random intercepts;
\code{"individual"}) penalty is created.}

\item{node_labels}{character; a vector of alternative labels for the levels
of the factor.}

\item{add_delta}{numeric or logical; either the numeric value to add to the
diagonal of the MRF penalty matrix, or a logical value indicating if such
an adjustment should be made. The default is to not alter the diagonal of
the penalty matrix.}

\item{...}{arguments passed to other methods.}
}
\description{
Fully connected graph and random effect MRF penalties from a factor
}
