% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties_tree.R
\name{mrf_penalty.phylo}
\alias{mrf_penalty.phylo}
\title{MRF penalty from a phylogeny}
\usage{
\method{mrf_penalty}{phylo}(
  object,
  model = c("rw1", "ou", "brownian"),
  alpha = NULL,
  at_tips = NULL,
  internal_nodes = TRUE,
  delta = FALSE,
  ...
)
}
\arguments{
\item{object}{an R object to create the MRF penalty from.}

\item{model}{character; one of \code{"rw1"}, \code{"ou"}, or \code{"brownian"}. The type of
trait evolution model to use.}

\item{alpha}{numeric; the autoregressive parameter for an OU stochastic
process. Should be >1e-5 (alpha = 0 would correspond to a "rw1" model).}

\item{at_tips}{character; vector of tip labels to calculate the penalty at.
All values in the vector must correpond to tip name labels in the tree.}

\item{internal_nodes}{logical; should the internal nodes of the tree be
included in the penalty (\code{TRUE}), or just the tips (terminal nodes;
\code{FALSE})}

\item{delta}{numeric or logical; either the numeric value to add to the
diagonal of the MRF penalty matrix, or a logical value indicating if such
an adjustment should be made. The default is to not alter the diagonal of
the penalty matrix.}

\item{...}{arguments passed to other methods.}
}
\description{
MRF penalty from a phylogeny
}
\examples{
#Example code

}
