% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{mrf_penalty}
\alias{mrf_penalty}
\title{Markov Random Field Penalty}
\usage{
mrf_penalty(object, ...)
}
\arguments{
\item{object}{an R object to create the MRF penalty from.}

\item{...}{arguments passed to other methods.}
}
\description{
Markov Random Field Penalty
}
