% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{prec_ar1}
\alias{prec_ar1}
\title{Generator function for a 1st order discrete-time autoregessive
process.}
\usage{
prec_ar1(start, end, n, dists, rho)
}
\arguments{
\item{start}{vector of starting indices. All values must be between 1 and n}

\item{end}{vector of ending indices (nodes that follow the start node in the
list)}

\item{n}{Total number of nodes in the system}

\item{dists}{The distances between the pairs of nodes. Must be positive
integers.}

\item{rho}{The autocorrelation strength of the random walk. Must range
between -1 and 1.}
}
\description{
Internal function that calculates the precision matrix for a 1d
1st order discrete-time autoregressive random walk. Takes two vectors of
indices (start and end) saying which nodes out of the n total nodes are
connected to each other. The \code{rho} parameter controls the strength of the
autoregressive component relative to the ends of function.
}
\keyword{internal}
