% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{prec_rw1}
\alias{prec_rw1}
\title{Generator for 1st order randm walk}
\usage{
prec_rw1(start, end, n, dists)
}
\arguments{
\item{start}{Indices for the starting value of an arrow connecting two obs}

\item{end}{Indices for the starting value of an arrow connecting two obs.
Must be the same length as \code{start}.}

\item{n}{Total number of values for the smoother}

\item{dists}{Distances between pairs of observations. Must be a vector of
positive values of the same length as \code{start}}
}
\description{
Generator for 1st order randm walk

Internal function for generating precision matrices for 1st order RWs
for 1D data. This is the same as a Brownian motion stochastic process
sampled at a finite number of points
}
\keyword{internal}
