% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{prec_rw2}
\alias{prec_rw2}
\title{Generator function for a 2nd order continuous-time random walk for
irregularly sampled points}
\usage{
prec_rw2(dists, n, end_dist = Inf, derivs = FALSE)
}
\arguments{
\item{dists}{The distances between the pairs of nodes. Must be positive
numbers}

\item{n}{The number of nodes in the series. Must be equal to 1 + the number
of distances}

\item{end_dist}{The distance between the end points. If infinite, the random
walk is non-cyclic.}

\item{derivs}{Whether to return the sparse matrix (with both derivatives and
values of the function) or the dense matrix for the function values only.}
}
\description{
Internal function that calculates the precision matrix for a
1-dimensional 2nd order continuous time random walk. The underlying model
is an integrated Wiener process (RW2 in Rue and Held, 2005) across possibly
irregularly spaced points. Values are assumed to be in order, with
distances indexing how far apart observations are. Can be made into a
cyclic RW2 model by specifying a finite end distance (\code{end_dist}).

The sparse version of this process gives a 2n x 2n block matrix; the upper
n x n block is the precision matrix for the values of the function at the
specified points, the lower right n x n block is the precision matrix for
the derivatives of the function at those points. It has rank 2n-2 (both the
flat line and the linear function fall in the null space).

The dense version of this precision matrix is an n x n matrix for only the
values of the function. It has rank n-1 (the flat line is in the null
space)
}
\keyword{internal}
