% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_func1D}
\alias{sim_func1D}
\title{Simulate 1D nonlinear functions}
\usage{
sim_func1D(
  n_obs = 100,
  n_funcs = 1,
  trend_type = c("linear", "none", "exponential", "saturating", "sigmoid"),
  discrete = FALSE,
  irregular = FALSE,
  n_sines = 5,
  spec_power = 1,
  cycle_strength = 1,
  trend_strength = 1,
  intercept_sd = 1,
  trend_scale = 1,
  derivs = FALSE
)
}
\description{
Simulates nonlinear functions across a range of values.
}
