% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{vis}
\alias{vis}
\alias{visualise}
\title{Visualize a data set or object}
\usage{
vis(x, ...)

visualise(x, ...)
}
\arguments{
\item{x}{A data frame or other object.}

\item{...}{Other arguments passed to methods}
}
\description{
Synonyms of the \code{\link[generics:visualize]{generics::visualize()}} method. Alternatives are \code{vis()} and
\code{visualise()}.
}
