% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{visualize.dendrogram_mrf_penalty}
\alias{visualize.dendrogram_mrf_penalty}
\title{Plot a thing}
\usage{
\method{visualize}{dendrogram_mrf_penalty}(
  x,
  graph = TRUE,
  layout = "stress",
  circular = FALSE,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  fill_scale = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"dendrogram_mrf_penalty"}}
}
\description{
Plot a thing
}
\examples{
# example code
hc <- hclust(dist(USArrests), "complete")
mrf_penalty(hc) |>
  visualize()
}
