% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{visualize.first_order_random_walk_mrf_penalty}
\alias{visualize.first_order_random_walk_mrf_penalty}
\title{Plot a thing}
\usage{
\method{visualize}{first_order_random_walk_mrf_penalty}(
  x,
  graph = TRUE,
  layout = "linear",
  circular = FALSE,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  fill_scale = NULL
)
}
\arguments{
\item{x}{an object of class \code{"first_order_random_walk_mrf_penalty"}}

\item{graph}{logical;}

\item{layout}{character;}

\item{circular}{logical;}
}
\description{
Plot a thing
}
\examples{
# example code
mrf_penalty(1:10, type = "linear") |>
  visualize()
}
