% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{zero_pad}
\alias{zero_pad}
\title{Add zeros onto numeric factor level names}
\usage{
zero_pad(x)
}
\description{
add leading zeros to a vector of numbers to improve sorting. Not
currently used in code; here if needed.
}
\keyword{internal}
