% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_type}
\alias{get_type}
\alias{get_type.mrf_penalty}
\alias{get_type.mrf_config}
\title{Extract the type of MRF from the penalty}
\usage{
get_type(object)

\method{get_type}{mrf_penalty}(object)

\method{get_type}{mrf_config}(object)
}
\arguments{
\item{object}{an object of class \code{"mrf_penalty"} or \code{"mrf_config"}.}
}
\value{
A length 1 character vector containing the type of MRF penalty.
}
\description{
Extract the type of MRF from the penalty
}
