% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{mrf_penalty.dendrogram}
\alias{mrf_penalty.dendrogram}
\title{MRF penalty from a dendrogram}
\usage{
\method{mrf_penalty}{dendrogram}(object, node_labels = NULL,
  add_delta = FALSE, ...)
}
\description{
MRF penalty from a dendrogram
}
