% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{mrf_penalty.dendrogram}
\alias{mrf_penalty.dendrogram}
\title{MRF penalty from a dendrogram}
\usage{
\method{mrf_penalty}{dendrogram}(object, node_labels = NULL, add_delta = FALSE, ...)
}
\arguments{
\item{object}{an R object to create the MRF penalty from.}

\item{node_labels}{character; a vector of alternative labels for the levels
of the factor.}

\item{add_delta}{numeric or logical; either the numeric value to add to the
diagonal of the MRF penalty matrix, or a logical value indicating if such
an adjustment should be made. The default is to not alter the diagonal of
the penalty matrix.}

\item{...}{arguments passed to other methods.}
}
\description{
MRF penalty from a dendrogram
}
