% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{mrf_penalty.hclust}
\alias{mrf_penalty.hclust}
\title{MRF penalty from a hclust object}
\usage{
\method{mrf_penalty}{hclust}(object, ...)
}
\arguments{
\item{object}{an R object to create the MRF penalty from.}

\item{...}{arguments passed to other methods.}
}
\description{
MRF penalty from a hclust object
}
