% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{mrf_penalty.numeric}
\alias{mrf_penalty.numeric}
\title{Continuous-time random walk MRF penalty from a numeric vector}
\usage{
\method{mrf_penalty}{numeric}(
  object,
  model = c("rw1", "rw2", "ar1", "ou"),
  cyclic = FALSE,
  rho = NULL,
  at_nodes = NULL,
  node_labels = NULL,
  add_missing = NULL,
  end_points = NULL,
  end_dist = NULL,
  delta = FALSE,
  ...
)
}
\arguments{
\item{object}{an R object to create the MRF penalty from.}

\item{model}{character; one of "rw1", "rw2", "ar1", or "ou". See Description
for details on the models}

\item{cyclic}{logical; If TRUE, the end points are treated as neighbouring
each other. See Description for details}

\item{node_labels}{character; a vector of alternative labels for the levels
of the factor.}

\item{end_points}{numeric; an optional vector of length 2 providing the end
points of the period of cycle.}

\item{delta}{numeric or logical; either the numeric value to add to the
diagonal of the MRF penalty matrix, or a logical value indicating if such
an adjustment should be made. The default is to not alter the diagonal of
the penalty matrix.}

\item{...}{arguments passed to other methods.}
}
\description{
Models one-dimensional numeric vectors as random-walk models. \strong{STUB}
}
\examples{
# rw1: 1st order continuous-time random walk
p <- mrf_penalty(1:10, model = "rw1")
as.matrix(p)

# cyclic rw1:
p <- mrf_penalty(1:10, model = "rw1", cyclic = TRUE)
as.matrix(p)

# cyclic with user-specified end points
p <- mrf_penalty(1:10, model = "rw1",  end_points = c(0,11))
as.matrix(p)
}
