% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{mrf_penalty.numeric}
\alias{mrf_penalty.numeric}
\title{First-order random walk MRF penalty from a numeric vector}
\usage{
\method{mrf_penalty}{numeric}(
  object,
  type = c("linear", "cyclic"),
  node_labels = NULL,
  add_delta = FALSE,
  end_points = NULL,
  ...
)
}
\arguments{
\item{object}{an R object to create the MRF penalty from.}

\item{type}{character; one of \code{"linear"} or \code{"cyclic"} indicating if the
observations form a cyclic series or not.}

\item{node_labels}{character; a vector of alternative labels for the levels
of the factor.}

\item{add_delta}{numeric or logical; either the numeric value to add to the
diagonal of the MRF penalty matrix, or a logical value indicating if such
an adjustment should be made. The default is to not alter the diagonal of
the penalty matrix.}

\item{end_points}{numeric; an optional vector of length 2 providing the
end points of the period of cycle.}

\item{...}{arguments passed to other methods.}
}
\description{
First-order random walk MRF penalty from a numeric vector
}
