% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{visualize.cyclic_mrf_penalty}
\alias{visualize.cyclic_mrf_penalty}
\title{Visualizing penalty matrix or graph object for a cyclic 1D MRF}
\usage{
\method{visualize}{cyclic_mrf_penalty}(x, graph = TRUE, layout = "linear", ...)
}
\arguments{
\item{x}{an object of class \code{"cyclic_mrf_penalty"}}

\item{graph}{logical;}

\item{layout}{character;}

\item{circular}{logical;}
}
\description{
Visualizing penalty matrix or graph object for a cyclic 1D MRF
}
\examples{
# example code
mrf_penalty(1:10, type = "linear") |>
  visualize()
}
