% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{visualize.first_order_random_walk_mrf_penalty}
\alias{visualize.first_order_random_walk_mrf_penalty}
\title{Plot a thing}
\usage{
\method{visualize}{first_order_random_walk_mrf_penalty}(
  x,
  graph = TRUE,
  layout = "linear",
  circular = FALSE,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  fill_scale = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"first_order_random_walk_mrf_penalty"}}

\item{graph}{logical;}

\item{layout}{character;}

\item{circular}{logical;}

\item{xlab, ylab, title, subtitle, caption}{character; labels for plots. If
\code{xlab} or \code{ylab} are not supplied, a suitable default is used.}

\item{fill_scale}{a suitable fill scale to use if plotting the penalty
matrix}

\item{...}{arguments passed to other methods}
}
\description{
Plot a thing
}
\examples{
# example code
mrf_penalty(1:10, type = "linear") |>
  visualize()
}
