% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{visualize.sequential_mrf_penalty}
\alias{visualize.sequential_mrf_penalty}
\title{Plot a thing}
\usage{
\method{visualize}{sequential_mrf_penalty}(x, graph = TRUE, layout = "linear", circular = FALSE)
}
\arguments{
\item{x}{an object of class \code{"sequential_mrf_penalty"}}

\item{graph}{logical;}

\item{layout}{character;}

\item{circular}{logical;}
}
\description{
Plot a thing
}
\examples{
# example code
mrf_penalty(1:10, type = "linear") |>
  visualize()
}
